refactor class component ImageGallery to functional component
