+++ ProductDetails/test/App.test.js
@@ -1,8 +1,7 @@
 // Link.react.test.js
 import React from 'react';
 import { shallow, mount, render } from 'enzyme';
-import App from '../src/components/Products.js';
-import SubApp from '../src/components/SubApp.js'
+import App from '../src/components/Products.jsx';
 
 it('Link changes the class when hovered', () => {
     const wrapper = shallow(<App />);
